cephfs admin: move tests that were previously only built on octopus
